requirements.txt
Flask
Flask-Cors
python-dotenv
streamlit
requests